LIKE support, DATETIME fix
Added support for LIKE in where clauses.
Fixed parsing of date/time strings in TupleUtil::setTupleField. The
date/time strings would match if only the first part of the format
parsed correctly. Reordered so that more complex formats are parsed
first before simpler formats.
Updated instructions in the beginning of the example files.